renderer: Refactor filenames for render stages.
