+++ libopenage/gamestate/terrain_chunk.h
@@ -6,7 +6,7 @@
 
 #include "coord/tile.h"
 #include "gamestate/terrain_tile.h"
-#include "renderer/stages/terrain/terrain_render_entity.h"
+#include "renderer/stages/terrain/render_entity.h"
 #include "time/time.h"
 #include "util/vector.h"
