+++ libopenage/presenter/presenter.cpp
@@ -25,11 +25,11 @@
 #include "renderer/resources/shader_source.h"
 #include "renderer/resources/texture_info.h"
 #include "renderer/stages/camera/manager.h"
-#include "renderer/stages/hud/hud_renderer.h"
-#include "renderer/stages/screen/screen_renderer.h"
-#include "renderer/stages/skybox/skybox_renderer.h"
-#include "renderer/stages/terrain/terrain_renderer.h"
-#include "renderer/stages/world/world_renderer.h"
+#include "renderer/stages/hud/render_stage.h"
+#include "renderer/stages/screen/render_stage.h"
+#include "renderer/stages/skybox/render_stage.h"
+#include "renderer/stages/terrain/render_stage.h"
+#include "renderer/stages/world/render_stage.h"
 #include "renderer/window.h"
 #include "time/time_loop.h"
 #include "util/path.h"
