+++ libopenage/renderer/demo/demo_3.cpp
@@ -13,12 +13,12 @@
 #include "renderer/resources/assets/asset_manager.h"
 #include "renderer/resources/shader_source.h"
 #include "renderer/stages/camera/manager.h"
-#include "renderer/stages/screen/screen_renderer.h"
-#include "renderer/stages/skybox/skybox_renderer.h"
-#include "renderer/stages/terrain/terrain_render_entity.h"
-#include "renderer/stages/terrain/terrain_renderer.h"
-#include "renderer/stages/world/world_render_entity.h"
-#include "renderer/stages/world/world_renderer.h"
+#include "renderer/stages/screen/render_stage.h"
+#include "renderer/stages/skybox/render_stage.h"
+#include "renderer/stages/terrain/render_entity.h"
+#include "renderer/stages/terrain/render_stage.h"
+#include "renderer/stages/world/render_entity.h"
+#include "renderer/stages/world/render_stage.h"
 #include "renderer/uniform_buffer.h"
 #include "time/clock.h"
