+++ libopenage/renderer/stages/hud/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_sources(libopenage
-	hud_object.cpp
-	hud_render_entity.cpp
-	hud_renderer.cpp
+	object.cpp
+	render_entity.cpp
+	render_stage.cpp
 )
