+++ libopenage/renderer/stages/hud/object.h
