+++ libopenage/renderer/stages/hud/render_entity.h
