+++ libopenage/renderer/stages/hud/render_stage.h
