+++ libopenage/renderer/stages/screen/render_stage.h
