+++ libopenage/renderer/stages/skybox/render_stage.h
