+++ libopenage/renderer/stages/terrain/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_sources(libopenage
-    terrain_chunk.cpp
-	terrain_mesh.cpp
-	terrain_model.cpp
-	terrain_render_entity.cpp
-	terrain_renderer.cpp
+    chunk.cpp
+	mesh.cpp
+	model.cpp
+	render_entity.cpp
+	render_stage.cpp
 )
