+++ libopenage/renderer/stages/terrain/mesh.h
