+++ libopenage/renderer/stages/terrain/model.h
