+++ libopenage/renderer/stages/terrain/render_entity.cpp
@@ -1,6 +1,6 @@
 // Copyright 2022-2023 the openage authors. See copying.md for legal info.
 
-#include "terrain_render_entity.h"
+#include "render_entity.h"
 
 #include <algorithm>
 #include <array>
