+++ libopenage/renderer/stages/terrain/render_entity.h
