+++ libopenage/renderer/stages/terrain/render_stage.h
