+++ libopenage/renderer/stages/world/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_sources(libopenage
-	world_object.cpp
-	world_render_entity.cpp
-	world_renderer.cpp
+	object.cpp
+	render_entity.cpp
+	render_stage.cpp
 )
