+++ libopenage/renderer/stages/world/object.h
