+++ libopenage/renderer/stages/world/render_entity.cpp
@@ -1,6 +1,6 @@
 // Copyright 2022-2023 the openage authors. See copying.md for legal info.
 
-#include "world_render_entity.h"
+#include "render_entity.h"
 
 #include <functional>
 #include <mutex>
