+++ libopenage/renderer/stages/world/render_entity.h
