+++ libopenage/renderer/stages/world/render_stage.h
